:wrench: move files to match with folder structure
